third_party/WebKit/LayoutTests/fast/workers/worker-onerror-06.html

<!DOCTYPE html>
<html>
<head>
    <script>
        window.isOnErrorTest = true;
    </script>
    <script src="../../resources/js-test.js"></script>
    <script src="resources/onerror-test.js"></script>
</head>
<body>
    <!-- This script's body will be used to build a Blob URL to use as a Worker. -->
    <script id="workerCode" type="text/plain">
        onerror = function(message, filename, lineno, colno, error) {
            postMessage({
                message: message,
                filename: filename,
                lineno: lineno,
                colno: colno,
                error: error
            });
            return false;
        }

        foo.bar = 0;
    </script>
    <script>
        description("This tests that returning true from a page-level handler suppresses the exception.");

        checkErrorEventInHandler({
            message: "Uncaught ReferenceError: foo is not defined",
            filename: "[blob: URL]",
            lineno: 13,
            colno: 9,
        }, true);
    </script>
</body>
</html>
