third_party/WebKit/LayoutTests/fast/workers/worker-onerror-06-expected.txt

This tests that returning true from a page-level handler suppresses the exception.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Page-level worker.onerror handler triggered:
PASS errorEvent.message is "Uncaught ReferenceError: foo is not defined"
PASS stripURL(errorEvent.filename) is "[blob: URL]"
PASS errorEvent.lineno is 13
PASS errorEvent.colno is 9
PASS errorEvent.error is null
PASS successfullyParsed is true

TEST COMPLETE
